<fastio_user>: Connect DPDK l2fwd with preprocess.py

+++ fastio_user/include/fastio_user/ipc.h
@@ -5,9 +5,16 @@
 #ifndef IPC_H
 #define IPC_H
 
+/**
+ * @brief init_uds_stream_cli
+ *
+ * @param socket_path
+ *
+ * @return
+ */
 int init_uds_stream_cli(const char* socket_path);
 
-void send_mbufs_data_uds(int socket, struct rte_mbuf** buf, uint16_t size_b,
+void send_mbufs_data_uds(int socket, struct rte_mbuf** buf, uint32_t size_b,
     uint16_t nb_mbuf, uint16_t tail_size);
 
 #endif /* !IPC_H */
